--- src/Microsoft.Kusto.ServiceLayer/DataSource/Intellisense/IIntellisenseClient.cs
@@ -1,15 +1,0 @@
-using Microsoft.Kusto.ServiceLayer.LanguageServices;
-using Microsoft.Kusto.ServiceLayer.LanguageServices.Contracts;
-using Microsoft.Kusto.ServiceLayer.Workspace.Contracts;
-
-namespace Microsoft.Kusto.ServiceLayer.DataSource.Intellisense
-{
-    public interface IIntellisenseClient
-    {
-        void UpdateDatabase(string databaseName);
-        ScriptFileMarker[] GetSemanticMarkers(ScriptParseInfo parseInfo, ScriptFile scriptFile, string queryText);
-        DefinitionResult GetDefinition(string queryText, int index, int startLine, int startColumn, bool throwOnError = false);
-        Hover GetHoverHelp(ScriptDocumentInfo scriptDocumentInfo, Position textPosition, bool throwOnError = false);
-        CompletionItem[] GetAutoCompleteSuggestions(ScriptDocumentInfo scriptDocumentInfo, Position textPosition, bool throwOnError = false);
-    }
-}
